requirements.txt
awsebcli==3.12.1
blessed==1.14.2
boto3==1.5.10
botocore==1.8.24
cement==2.8.2
click==6.7
colorama==0.3.7
coverage==4.4.2
decorator==4.1.2
dj-database-url==0.4.2
Django==2.0.1
django-cors-headers==2.1.0
django-smtp-ssl==1.0
django-storages==1.6.5
djangorestframework==3.7.7
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
jmespath==0.9.3
parso==0.1.1
pathspec==0.5.0
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
python-dotenv==0.7.1
pytz==2017.3
PyYAML==3.12
requests==2.9.1
s3transfer==0.1.12
semantic-version==2.5.0
simplegeneric==0.8.1
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
traitlets==4.3.2
wcwidth==0.1.7
websocket-client==0.46.0
